Moving raw interface declarations into ingestion layer
